pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flippr"
dynamic = ["version"]
description = "FragPipe Limited-Proteolysis Processor (FLiPPR) is modular, fast, and easy-to-use data processing tool for LiP-MS & LFQ-based experiments analyzed in FragPipe"
readme = "README.md"
requires-python = ">=3.10"
keywords = [
  "bottom-up proteomics",
  "computational mass spectrometry",
  "fragpipe",
  "fragpipe analysis",
  "limited proteolysis",
  "lip-ms",
  "lfq",
  "label-free",
  "label-free quantification",
  "differential analysis",
  "proteomics",
  "flippr",
  "fragpipe"
]
authors = [
  { name = "Edgar Manriquez-Sandoval", email = "emanriq1@jhu.edu"},
]
maintainers = [
  {name = "FriedLab@JHU", email = "friedlab.jhu@gmail.com"},
  {name = "Stephen D. Fried", email = "sdfried@jhu.edu"},
  {name = "Edgar Manriquez-Sandoval", email = "emanriq1@jhu.edu"},
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "biopython",
  "h2",
  "httpx",
  "matplotlib",
  "nest_asyncio",
  "numpy",
  "pandas",
  "polars",
  "pyarrow",
  "requests",
  "scipy",
  "seaborn",
  "xlsx2csv",
  "xlsxwriter"
]

[project.urls]
Documentation = "https://github.com/FriedLabJHU/FragPipe-Limited-Proteolysis-Processor#readme"
Issues = "https://github.com/FriedLabJHU/FragPipe-Limited-Proteolysis-Processorissues"
Source = "https://github.com/FriedLabJHU/FragPipe-Limited-Proteolysis-Processor"

[tool.hatch.version]
path = "src/flippr/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[tool.setuptools.package-data]
"flippr" = ["py.typed"]

[[tool.hatch.envs.all.matrix]]
python = ["3.10", "3.11", "3.12"]

[tool.hatch.envs.types]
dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/flippr tests}"

[tool.coverage.run]
source_pkgs = ["flippr", "tests"]
branch = true
parallel = true
omit = [
  "src/flippr/__about__.py",
]

[tool.coverage.paths]
flippr = ["src/flippr", "*/flippr/src/flippr"]
tests = ["tests", "*/flippr/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
